# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuario-rol": {
            "get": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Lista todas las relaciones usuario-rol",
                "operationId": "cd05632ab9f6ff267317ac9e2ae142b0",
                "responses": {
                    "200": {
                        "description": "Lista de relaciones usuario-rol"
                    }
                }
            },
            "post": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Crear una nueva relación usuario-rol",
                "operationId": "045f9c74fc708e2b0331b4b0d0029147",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "rol_id"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer"
                                    },
                                    "rol_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación creada exitosamente"
                    }
                }
            }
        },
        "/api/usuario-rol/{id}": {
            "get": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Mostrar una relación usuario-rol específica",
                "operationId": "33b473d22eb02e30420e72b3166edf03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación usuario-rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación encontrada"
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Actualizar una relación usuario-rol existente",
                "operationId": "9352eb1c913840ac755402d6cf9292bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación usuario-rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "rol_id"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer"
                                    },
                                    "rol_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Relación actualizada correctamente"
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Eliminar una relación usuario-rol",
                "operationId": "2f5bcc29a09d06f841f0023be9fd7e4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación usuario-rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación eliminada correctamente"
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            }
        },
...
    },
...
}